+++ mrv2/lib/mrvFl/mrvFileRequester.cpp
@@ -355,12 +355,15 @@ namespace mrv
             ("Audios (*.{") + getAudioPattern(context) + "})\t";
         const std::string kIMAGE_PATTERN =
             _("Images (*.{") + getImagePattern(context) + "})\t";
+        const std::string kSESSIONS_PATTERN =
+            _("Sessions (*.") + kSessionPattern + "})\t";
         const std::string kALL_PATTERN =
             _("All (*.{") + getImagePattern(context) + "," +
             getMoviePattern(context) + "," + kReelPattern + "," +
-            getAudioPattern(context) + "," + kReelPattern + "})\t" +
+            getAudioPattern(context) + "," + kSessionPattern + "})\t" +
             kIMAGE_PATTERN + kAUDIO_PATTERN + _("Movies (*.{") +
-            getMoviePattern(context) + "})\t" + kREEL_PATTERN;
+            getMoviePattern(context) + "})\t" + kREEL_PATTERN +
+            kSESSIONS_PATTERN;
 
         std::string pattern = kIMAGE_PATTERN + kAUDIO_PATTERN;
         std::string title = _("Load Image(s)");